+++ CHANGELOG.md
@@ -16,3 +16,10 @@
 * Handle change to Sectigo API for organization/department lookups
 # 1.1.0
 * Allow organization name to be provided in the template section of the config
+# 1.2.0
+* Allow for blank CN to be provided
+* Fixes for Certificate Authentication  
+# 1.2.1  
+* Fix for handling sync of expired records  
+* Handle null Keyfactor-Requestor  
+* Properly pass in custom fields to enrollment API